--- packages/firebase_ui_firestore/example/test_driver/firebase_ui_firestore_e2e_test.dart
@@ -1,7 +1,0 @@
-// Copyright 2020, the Chromium project authors.  Please see the AUTHORS file
-// for details. All rights reserved. Use of this source code is governed by a
-// BSD-style license that can be found in the LICENSE file.
-
-import 'package:drive/drive_driver.dart' as drive;
-
-void main() => drive.main();
