+++ tests/integration_test/firebase_ui_auth/firebase_options.dart
