+++ tests/integration_test/firebase_ui_auth/phone_verification_test.dart
@@ -9,7 +9,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 import 'package:firebase_ui_localizations/firebase_ui_localizations.dart';
 
-import 'utils.dart';
+import '../utils.dart';
 
 Future<void> sendSMS(WidgetTester tester, String phoneNumber) async {
   await tester.pump();
