+++ tests/integration_test/firebase_ui_firestore/firebase_ui_firestore_e2e.dart
@@ -3,15 +3,13 @@
 // BSD-style license that can be found in the LICENSE file.
 
 import 'package:flutter_test/flutter_test.dart';
-import 'package:integration_test/integration_test.dart';
 
 import 'firestore_list_view_test.dart' as firestore_list_view_test;
 import 'firestore_query_builder_test.dart' as firestore_query_builder_test;
 
-import 'utils.dart';
+import '../utils.dart';
 
 Future<void> main() async {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
   setUpAll(prepare);
 
   firestore_list_view_test.main();
