+++ tests/integration_test/firebase_ui_firestore/firestore_query_builder_test.dart
@@ -8,7 +8,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:firebase_ui_firestore/firebase_ui_firestore.dart';
 import 'package:mockito/mockito.dart';
-import 'utils.dart';
+import '../utils.dart';
 
 void main() {
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
