+++ tests/integration_test/firebase_ui_oauth_facebook/facebook_sign_in_test.dart
@@ -2,38 +2,33 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
+import 'package:flutter_facebook_auth/flutter_facebook_auth.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 import 'package:firebase_ui_localizations/firebase_ui_localizations.dart';
 import 'package:firebase_ui_oauth/firebase_ui_oauth.dart';
-import 'package:firebase_ui_oauth_twitter/firebase_ui_oauth_twitter.dart';
+import 'package:firebase_ui_oauth_facebook/firebase_ui_oauth_facebook.dart';
 import 'package:mockito/mockito.dart';
-import 'package:twitter_login/twitter_login.dart';
-import 'package:twitter_login/entity/auth_result.dart' as twe;
 
-import 'utils.dart';
+import '../utils.dart';
 
 void main() async {
-  late TwitterProvider provider = TwitterProvider(
-    apiKey: 'apiKey',
-    apiSecretKey: 'apiSecretKey',
-  );
+  late FacebookProvider provider = FacebookProvider(clientId: 'clientId');
 
   setUp(() {
-    provider.provider = MockTwitterLogin();
+    provider.provider = MockFacebookAuth();
   });
 
   const labels = DefaultLocalizations();
 
   group(
-    'Sign in with Twitter button',
+    'Sign in with Facebook button',
     () {
       testWidgets('has a correct button label', (tester) async {
         await render(tester, OAuthProviderButton(provider: provider));
-        expect(find.text(labels.signInWithTwitterButtonText), findsOneWidget);
+        expect(find.text(labels.signInWithFacebookButtonText), findsOneWidget);
       });
 
       testWidgets(
@@ -65,7 +60,7 @@ void main() async {
           when(provider.provider.login()).thenAnswer(
             (realInvocation) async {
               await Future.delayed(const Duration(milliseconds: 50));
-              return MockAuthResult();
+              return MockLoginResult();
             },
           );
 
@@ -87,7 +82,7 @@ void main() async {
         await tester.tap(button);
         await tester.pumpAndSettle();
 
-        final user = FirebaseAuth.instance.currentUser!;
+        final user = auth.currentUser!;
 
         expect(user.displayName, 'Test User');
         expect(user.email, 'test@test.com');
@@ -107,25 +102,31 @@ void main() async {
 const _jwt =
     'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IlRlc3QgVXNlciIsImVtYWlsIjoidGVzdEB0ZXN0LmNvbSIsImlhdCI6MTUxNjIzOTAyMn0.m5qYto_Vs5ELTURC8rkD-JAJuoosdQZeuUZ_qFrEiaE';
 
-class MockAuthResult extends Mock implements twe.AuthResult {
+class MockAccessToken extends Mock implements AccessToken {
   @override
-  TwitterLoginStatus? get status => TwitterLoginStatus.loggedIn;
-  @override
-  String? get authToken => _jwt;
-  @override
-  String? get authTokenSecret => 'secret';
+  String get token => _jwt;
 }
 
-class MockTwitterLogin extends Mock implements TwitterLogin {
+class MockLoginResult extends Mock implements LoginResult {
   @override
-  Future<twe.AuthResult> login({bool? forceLogin}) async {
+  LoginStatus get status => LoginStatus.success;
+  @override
+  AccessToken? get accessToken => MockAccessToken();
+}
+
+class MockFacebookAuth extends Mock implements FacebookAuth {
+  @override
+  Future<LoginResult> login({
+    List<String>? permissions = const ['email', 'public_profile'],
+    LoginBehavior? loginBehavior = LoginBehavior.dialogOnly,
+  }) async {
     return super.noSuchMethod(
-      Invocation.method(
-        #signIn,
-        [],
-      ),
-      returnValue: MockAuthResult(),
-      returnValueForMissingStub: MockAuthResult(),
+      Invocation.method(#signIn, [], {
+        #permissions: permissions,
+        #behavior: loginBehavior,
+      }),
+      returnValue: MockLoginResult(),
+      returnValueForMissingStub: MockLoginResult(),
     );
   }
 }
