+++ tests/integration_test/firebase_ui_oauth_twitter/twitter_sign_in_test.dart
@@ -2,38 +2,37 @@
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
-import 'package:firebase_auth/firebase_auth.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:firebase_ui_auth/firebase_ui_auth.dart';
 import 'package:firebase_ui_localizations/firebase_ui_localizations.dart';
 import 'package:firebase_ui_oauth/firebase_ui_oauth.dart';
-import 'package:firebase_ui_oauth_google/firebase_ui_oauth_google.dart';
-import 'package:google_sign_in/google_sign_in.dart';
+import 'package:firebase_ui_oauth_twitter/firebase_ui_oauth_twitter.dart';
 import 'package:mockito/mockito.dart';
+import 'package:twitter_login/twitter_login.dart';
+import 'package:twitter_login/entity/auth_result.dart' as twe;
 
-import 'utils.dart';
+import '../utils.dart';
 
 void main() async {
-  late GoogleProvider provider = GoogleProvider(
-    clientId: 'clientId',
-    redirectUri: 'redirectUri',
-    scopes: const ['scope1', 'scope2'],
+  late TwitterProvider provider = TwitterProvider(
+    apiKey: 'apiKey',
+    apiSecretKey: 'apiSecretKey',
   );
 
   setUp(() {
-    provider.provider = MockGoogleSignIn();
+    provider.provider = MockTwitterLogin();
   });
 
   const labels = DefaultLocalizations();
 
   group(
-    'Sign in with Google button',
+    'Sign in with Twitter button',
     () {
       testWidgets('has a correct button label', (tester) async {
         await render(tester, OAuthProviderButton(provider: provider));
-        expect(find.text(labels.signInWithGoogleButtonText), findsOneWidget);
+        expect(find.text(labels.signInWithTwitterButtonText), findsOneWidget);
       });
 
       testWidgets(
@@ -48,7 +47,7 @@ void main() async {
           await tester.tap(button);
 
           await tester.pumpAndSettle();
-          verify(provider.provider.signIn()).called(1);
+          verify(provider.provider.login()).called(1);
 
           expect(true, isTrue);
         },
@@ -62,10 +61,10 @@ void main() async {
             OAuthProviderButton(provider: provider),
           );
 
-          when(provider.provider.signIn()).thenAnswer(
+          when(provider.provider.login()).thenAnswer(
             (realInvocation) async {
               await Future.delayed(const Duration(milliseconds: 50));
-              return MockGoogleSignInAccount();
+              return MockAuthResult();
             },
           );
 
@@ -87,24 +86,10 @@ void main() async {
         await tester.tap(button);
         await tester.pumpAndSettle();
 
-        final user = FirebaseAuth.instance.currentUser!;
+        final user = auth.currentUser!;
 
         expect(user.displayName, 'Test User');
         expect(user.email, 'test@test.com');
-      });
-
-      testWidgets('works standalone', (tester) async {
-        await render(
-          tester,
-          const GoogleSignInButton(
-            loadingIndicator: CircularProgressIndicator(),
-            clientId: 'test',
-          ),
-        );
-
-        final button = find.byType(GoogleSignInButton);
-        await tester.tap(button);
-        await tester.pump();
       });
     },
     skip: !provider.supportsPlatform(defaultTargetPlatform),
@@ -121,25 +106,25 @@ void main() async {
 const _jwt =
     'eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJzdWIiOiIxMjM0NTY3ODkwIiwibmFtZSI6IlRlc3QgVXNlciIsImVtYWlsIjoidGVzdEB0ZXN0LmNvbSIsImlhdCI6MTUxNjIzOTAyMn0.m5qYto_Vs5ELTURC8rkD-JAJuoosdQZeuUZ_qFrEiaE';
 
-class MockAuthentication extends Mock implements GoogleSignInAuthentication {
+class MockAuthResult extends Mock implements twe.AuthResult {
   @override
-  final String accessToken = _jwt;
+  TwitterLoginStatus? get status => TwitterLoginStatus.loggedIn;
+  @override
+  String? get authToken => _jwt;
+  @override
+  String? get authTokenSecret => 'secret';
 }
 
-// ignore: avoid_implementing_value_types, must_be_immutable
-class MockGoogleSignInAccount extends Mock implements GoogleSignInAccount {
+class MockTwitterLogin extends Mock implements TwitterLogin {
   @override
-  Future<GoogleSignInAuthentication> get authentication async =>
-      MockAuthentication();
-}
-
-class MockGoogleSignIn extends Mock implements GoogleSignIn {
-  @override
-  Future<GoogleSignInAccount?> signIn() async {
+  Future<twe.AuthResult> login({bool? forceLogin}) async {
     return super.noSuchMethod(
-      Invocation.method(#signIn, []),
-      returnValue: MockGoogleSignInAccount(),
-      returnValueForMissingStub: MockGoogleSignInAccount(),
+      Invocation.method(
+        #signIn,
+        [],
+      ),
+      returnValue: MockAuthResult(),
+      returnValueForMissingStub: MockAuthResult(),
     );
   }
 }
