+++ tests/ios/firebase_app_id_file.json
@@ -1,0 +1,7 @@
+{
+  "file_generated_by": "FlutterFire CLI",
+  "purpose": "FirebaseAppID & ProjectID for this Firebase app in this directory",
+  "GOOGLE_APP_ID": "1:406099696497:ios:acd9c8e17b5e620e3574d0",
+  "FIREBASE_PROJECT_ID": "flutterfire-e2e-tests",
+  "GCM_SENDER_ID": "406099696497"
+}
