+++ tests/linux/flutter/generated_plugin_registrant.cc
@@ -6,6 +6,14 @@
 
 #include "generated_plugin_registrant.h"
 
+#include <desktop_webview_auth/desktop_webview_auth_plugin.h>
+#include <flutter_secure_storage_linux/flutter_secure_storage_linux_plugin.h>
 
 void fl_register_plugins(FlPluginRegistry* registry) {
+  g_autoptr(FlPluginRegistrar) desktop_webview_auth_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "DesktopWebviewAuthPlugin");
+  desktop_webview_auth_plugin_register_with_registrar(desktop_webview_auth_registrar);
+  g_autoptr(FlPluginRegistrar) flutter_secure_storage_linux_registrar =
+      fl_plugin_registry_get_registrar_for_plugin(registry, "FlutterSecureStorageLinuxPlugin");
+  flutter_secure_storage_linux_plugin_register_with_registrar(flutter_secure_storage_linux_registrar);
 }
